// src/gradio/types.ts
export interface Config {
    auth_required?: true
    analytics_enabled: boolean
    connect_heartbeat: boolean
    auth_message: string
    components: ComponentMeta[]
    css: string | null
    js: string | null
    head: string | null
    dependencies: Dependency[]
    dev_mode: boolean
    enable_queue: boolean
    show_error: boolean
    layout: any
    mode: 'blocks' | 'interface'
    root: string
    root_url?: string
    theme: string
    title: string
    version: string
    space_id: string | null
    is_space: boolean
    is_colab: boolean
    show_api: boolean
    stylesheets: string[]
    path: string
    protocol: 'sse_v3' | 'sse_v2.1' | 'sse_v2' | 'sse_v1' | 'sse' | 'ws'
    max_file_size?: number
    theme_hash?: number
    username: string | null
}

// todo: DRY up types
export interface ComponentMeta {
    type: string
    id: number
    has_modes: boolean
    props: SharedProps
    instance?: any
    component?: any
    documentation?: Documentation
    children?: ComponentMeta[]
    parent?: ComponentMeta
    value?: any
    component_class_id: string
    key: string | number | null
    rendered_in?: number
}

interface SharedProps {
    elem_id?: string
    elem_classes?: string[]
    components?: string[]
    server_fns?: string[]
    interactive: boolean
    [key: string]: unknown
    root_url?: string
}

export interface Documentation {
    type?: TypeDescription
    description?: TypeDescription
    example_data?: string
}

interface TypeDescription {
    input_payload?: string
    response_object?: string
    payload?: string
}

export interface Dependency {
    id: number
    targets: [number, string][]
    inputs: number[]
    outputs: number[]
    backend_fn: boolean
    js: string | null
    scroll_to_output: boolean
    trigger: 'click' | 'load' | string
    max_batch_size: number
    show_progress: 'full' | 'minimal' | 'hidden'
    frontend_fn: ((...args: unknown[]) => Promise<unknown[]>) | null
    status?: string
    queue: boolean | null
    every: number | null
    batch: boolean
    api_name: string | null
    cancels: number[]
    types: DependencyTypes
    collects_event_data: boolean
    pending_request?: boolean
    trigger_after?: number
    trigger_only_on_success?: boolean
    trigger_mode: 'once' | 'multiple' | 'always_last'
    final_event: Payload | null
    show_api: boolean
    zerogpu?: boolean
    rendered_in: number | null
}

export interface DependencyTypes {
    generator: boolean
    cancel: boolean
}

export interface Payload {
    fn_index: number
    data: unknown[]
    time?: Date
    event_data?: unknown
    trigger_id?: number | null
}